adapted package structure to more general one, to separate thingiverse and facebook more.
