+++ src/com/t_oster/visicut/gui/MainView.form
@@ -952,7 +952,7 @@
     <Component class="javax.swing.JButton" name="btThingiverse">
       <Properties>
         <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
-          <Connection code="ThingiversePlatformIcon.get(ThingiversePlatformIcon.THINGIVERSE_LOGO)" type="code"/>
+          <Connection code="com.tur0kk.SocialPlatformIcon.get(com.tur0kk.SocialPlatformIcon.THINGIVERSE_LOGO)" type="code"/>
         </Property>
         <Property name="text" type="java.lang.String" resourceKey="btThingiverse.text"/>
         <Property name="toolTipText" type="java.lang.String" resourceKey="btThingiverse.toolTipText"/>
@@ -968,7 +968,7 @@
     <Component class="javax.swing.JButton" name="btFacebook">
       <Properties>
         <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
-          <Connection code="ThingiversePlatformIcon.get(ThingiversePlatformIcon.FACEBOOK_LOGO)" type="code"/>
+          <Connection code="com.tur0kk.SocialPlatformIcon.get(com.tur0kk.SocialPlatformIcon.FACEBOOK_LOGO)" type="code"/>
         </Property>
         <Property name="text" type="java.lang.String" resourceKey="btFacebook.text"/>
         <Property name="toolTipText" type="java.lang.String" resourceKey="btFacebook.toolTipText"/>
