+++ src/com/t_oster/visicut/gui/MainView.java
@@ -64,7 +64,7 @@ import com.t_oster.visicut.model.mapping.MappingSet;
 import com.tur0kk.facebook.FacebookManager;
 import com.tur0kk.facebook.gui.FacebookDialog;
 import com.tur0kk.thingiverse.ThingiverseManager;
-import com.tur0kk.thingiverse.uicomponents.ThingiversePlatformIcon;
+import com.tur0kk.SocialPlatformIcon;
 import java.awt.Desktop;
 import java.awt.Dimension;
 import java.awt.FileDialog;
@@ -1010,7 +1010,7 @@ public class MainView extends javax.swing.JFrame
         warningPanel.setName("warningPanel"); // NOI18N
         warningPanel.setPreferredSize(new java.awt.Dimension(276, 123));
 
-        btThingiverse.setIcon(ThingiversePlatformIcon.get(ThingiversePlatformIcon.THINGIVERSE_LOGO));
+        btThingiverse.setIcon(SocialPlatformIcon.get(SocialPlatformIcon.THINGIVERSE_LOGO));
         btThingiverse.setText(resourceMap.getString("btThingiverse.text")); // NOI18N
         btThingiverse.setToolTipText(resourceMap.getString("btThingiverse.toolTipText")); // NOI18N
         btThingiverse.setName("btThingiverse"); // NOI18N
@@ -1020,7 +1020,7 @@ public class MainView extends javax.swing.JFrame
             }
         });
 
-        btFacebook.setIcon(ThingiversePlatformIcon.get(ThingiversePlatformIcon.FACEBOOK_LOGO));
+        btFacebook.setIcon(SocialPlatformIcon.get(SocialPlatformIcon.FACEBOOK_LOGO));
         btFacebook.setText(resourceMap.getString("btFacebook.text")); // NOI18N
         btFacebook.setToolTipText(resourceMap.getString("btFacebook.toolTipText")); // NOI18N
         btFacebook.setName("btFacebook"); // NOI18N
